installed react-timeline, broke progress bar into subcomponent, inserted example timeline
